Update skill-icons styles in style.css
In the last few commits, I was editing the <i>'s height and width. This did not work as it's an inline font element. Updating h/w changed the box size, but the font size would still reduce on mobile screens, leaving empty spaces in the box. It's better to use font-size to adjust the size of icons. So, I have updated the styles accordingly.